pyproject.toml
[project]
name = "enhanced-rag-system"
version = "0.1.0"
description = "Production-ready RAG system with agentic workflows, semantic caching, and hallucination detection"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    # Core LlamaIndex dependencies
    "python-dotenv>=1.0.0,<2.0.0",
    "pydantic>=2.11.5",
    "aiostream>=0.5.2,<0.6.0",
    "llama-index-core>=0.12.28,<0.13.0",
    "llama-index-readers-file>=0.4.6,<1.0.0",
    "llama-index-llms-openai>=0.3.2,<0.4.0",
    "llama-index-embeddings-openai>=0.3.1,<0.4.0",
    "llama-deploy",
    "docx2txt>=0.8,<0.9",
    
    # Enhanced retrieval and search
    "llama-index-retrievers-bm25>=0.4.0",
    "rank-bm25>=0.2.2",
    "nltk>=3.8.1",
    "sentence-transformers>=2.2.2",
    "torch>=2.0.0",
    
    # Semantic caching
    "redis>=4.0.0",
    "numpy>=1.21.0",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "mypy>=1.8.0,<2.0.0",
    "pytest>=8.3.5,<9.0.0",
    "pytest-asyncio>=0.25.3,<0.26.0"
]

# Multimodal features  
multimodal = [
    "transformers>=4.21.0",
    "pillow>=9.0.0",
    "opencv-python>=4.8.0", 
    "pdf2image>=1.16.0",
    "pytesseract>=0.3.10",
]

# Text-to-speech features
tts = [
    "pyttsx3>=2.90",
    "gTTS>=2.3.0",
]

# Advanced reranking (optional)
reranking = [
    "llama-index-postprocessor-cohere-rerank>=0.3.0",
]

# All optional features
all = [
    "enhanced-rag-system[multimodal,tts,reranking]"
]

[project.scripts]
generate = "src.generate:generate_index"

[tool.uv.sources]
llama-deploy = { git = "https://github.com/run-llama/llama_deploy" }

[tool.mypy]
python_version = "3.11"
plugins = "pydantic.mypy"
exclude = ["tests", "venv", ".venv", "output", "config"]
check_untyped_defs = true
warn_unused_ignores = false
show_error_codes = true
namespace_packages = true
ignore_missing_imports = true
follow_imports = "silent"
implicit_optional = true
strict_optional = false
disable_error_code = ["return-value", "assignment"]

[[tool.mypy.overrides]]
module = "src.*"
ignore_missing_imports = false

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"